auth-service/requirements.txt
Flask
PyJWT
flask-cors
pymongo
